backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pdfplumber
python-multipart
